Factor our tags and cigar strings into separate files for better reuse outside GFA
