+++ io/gfa.h
@@ -17,6 +17,8 @@
 
 #pragma once
 
+#include "cigar.h"
+
 #include <string_view>
 #include <optional>
 #include <string>
@@ -31,36 +33,10 @@
 #include <cinttypes>
 
 namespace gfa {
-struct tag {
-    char name[2];
-    char type;
-    std::variant<int64_t, std::string, float> val;
-
-    template<typename T>
-    tag(std::string_view n, std::string_view t, T v)
-            : name{n[0], n[1]}, type(t.front()), val(std::move(v)) {}
-
-    friend std::ostream &operator<<(std::ostream &s, const tag &t);
-
-    void print() const {
-        std::fprintf(stdout, "%c%c", name[0], name[1]);
-        std::fputs(":", stdout);
-        std::visit([&](const auto& value) { _print(value); }, val);
-    }
-
-  private:
-    void _print(int64_t i) const {
-        std::fprintf(stdout, "%c:%" PRId64, type, i);
-    }
-
-    void _print(const std::string &str) const {
-        std::fprintf(stdout, "%c:%s", type, str.c_str());
-    }
-
-    void _print(float f) const {
-        std::fprintf(stdout, "%c:%g", type, f);
-    }
-};
+using cigar::tag;
+using cigar::cigarop;
+using cigar::getTag;
+using cigar::cigar_string;
 
 struct header {
     std::vector<gfa::tag> tags;
@@ -101,17 +77,6 @@ struct segment {
         }
     }
 };
-
-struct cigarop {
-    uint32_t count : 24;
-    char op : 8;
-
-    void print() const {
-        std::fprintf(stdout, "%u%c", count, op);
-    }
-};
-
-using cigar_string = std::vector<cigarop>;
 
 struct link {
     std::string_view lhs;
@@ -229,34 +194,6 @@ struct path {
 };
 
 using record = std::variant<header, segment, link, gaplink, path>;
-
-static inline std::optional<gfa::tag>
-getTag(const char *name,
-       const std::vector<gfa::tag> &tags) {
-    auto res = std::find_if(tags.begin(), tags.end(),
-                            [=](const gfa::tag &tag) {
-                                return (tag.name[0] == name[0] &&
-                                        tag.name[1] == name[1]);
-                            });
-    if (res == tags.end())
-        return {};
-
-    return *res;
-}
-
-template<class T>
-std::optional<T> getTag(const char *name,
-                        const std::vector<gfa::tag> &tags) {
-    auto res = std::find_if(tags.begin(), tags.end(),
-                            [=](const gfa::tag &tag) {
-                                return (tag.name[0] == name[0] &&
-                                        tag.name[1] == name[1]);
-                            });
-    if (res == tags.end())
-        return {};
-
-    return std::get<T>(res->val);
-}
 
 std::optional<gfa::record> parseRecord(const char* line, size_t len);
 } // namespace gfa
